rename httpClient and fix tests
