+++ tests/test_http_client.py
@@ -1,6 +1,6 @@
 import unittest
 from importlib import reload
-from opencloning import httpClient
+from opencloning import http_client
 from opencloning import app_settings
 from pytest import MonkeyPatch
 from unittest.mock import patch, AsyncMock
@@ -14,15 +14,17 @@ class TestHttpClientProxy(unittest.IsolatedAsyncioTestCase):
     def tearDown(self):
         MonkeyPatch().delenv('PROXY_URL', raising=False)
         MonkeyPatch().delenv('PROXY_CERT_FILE', raising=False)
+        reload(app_settings)
+        reload(http_client)
 
     async def test_proxy(self):
-        MonkeyPatch().setenv('PROXY_URL', 'proxy_url')
+        MonkeyPatch().setenv('PROXY_URL', 'https://dummy.com')
         MonkeyPatch().setenv('PROXY_CERT_FILE', mock_cert_file)
         reload(app_settings)
-        reload(httpClient)
+        reload(http_client)
         mock_client = AsyncMock()
-        with patch('httpx.AsyncClient', return_value=mock_client) as mock_client_class:
-            async with httpClient.get_http_client():
+        with patch('opencloning.http_client.AsyncClient', return_value=mock_client) as mock_client_class:
+            async with http_client.get_http_client():
                 pass
 
             # Verify the client was initialized with correct proxy settings
@@ -30,11 +32,11 @@ class TestHttpClientProxy(unittest.IsolatedAsyncioTestCase):
             called_kwargs = mock_client_class.call_args.kwargs
             self.assertIn('proxy', called_kwargs)
             self.assertIn('verify', called_kwargs)
-            self.assertEqual(called_kwargs['proxy'], 'proxy_url')
+            self.assertEqual(called_kwargs['proxy'], 'https://dummy.com')
 
         # To ensure that the certificate is actually being loaded,
         # provide invalid path and test the error
         MonkeyPatch().setenv('PROXY_CERT_FILE', 'invalid_path')
         reload(app_settings)
         with self.assertRaises(FileNotFoundError):
-            reload(httpClient)
+            reload(http_client)
